flask_server/requirements.txt
flask
ultralytics
torch
gdown
Pillow
